feat: [C4R-1030] logic to handle fileManagement, working on be
