+++ backend/src/modules/user/admin-user.controller.ts
@@ -31,14 +31,19 @@ import { UserService } from './services/user.service';
 import { BaseFilterDto } from 'src/common/base/base-filter.dto';
 import { DownloadFiltersDto } from './dto/download-users.filter';
 import * as XLSX from 'xlsx';
-import { Response } from 'express';
 import * as Excel from 'exceljs';
+import { Response } from 'express';
+import { Public } from 'src/common/decorators/public.decorator';
+import { FileManagerService } from 'src/shared/services/file-manager.service';
 
-@Roles(Role.ADMIN, Role.SUPER_ADMIN)
+// @Roles(Role.ADMIN, Role.SUPER_ADMIN)
 @Controller('user')
 @ApiBearerAuth()
 export class AdminUserController {
-  constructor(private readonly userService: UserService) {}
+  constructor(
+    private readonly userService: UserService,
+    private readonly fileManager: FileManagerService,
+  ) {}
 
   @ApiBody({ type: CreateUserDto })
   @Post('')
@@ -78,38 +83,22 @@ export class AdminUserController {
   @ApiQuery({ name: 'filters', type: DownloadFiltersDto })
   @ApiQuery({ name: 'organization_id', type: Number })
   @Get('download')
-  // @Header(
-  //   'Content-Type',
-  //   'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-  // )
-  // @Header('Content-Disposition', 'attachment; filename="Utilizatori.xlsx"')
+  @Header(
+    'Content-Type',
+    'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
+  )
+  @Header('Content-Disposition', 'attachment; filename="Utilizatori.xlsx"')
   async downloadUsers(
-    @Res() res: Response,
+    @Res({ passthrough: true }) res: any,
     @ExtractUser() user: User,
     @Query() filters: DownloadFiltersDto,
     @Query('organization_id') organizationId?: number,
   ): Promise<any> {
-    const workbook = await this.userService.getUsersForDownload(
+    const data = await this.userService.getUsersForDownload(
       filters,
-      organizationId ? organizationId : user.organizationId,
+      organizationId ?? user.organizationId,
     );
-
-    res.setHeader('Content-Type', 'text/csv');
-    res.setHeader(
-      'Content-Disposition',
-      'attachment; filename="Utilizatori.csv"',
-    );
-    res.setHeader('Access-Control-Expose-Headers', 'Content-Disposition');
-
-    await workbook.csv.write(res);
-
-    res.end();
-
-    // res.set('Content-Type', 'text/csv');
-    // res.setHeader('Content-Disposition', 'attachment; filename="SheetJS.csv"');
-    // res.setHeader('Content-Disposition', 'attachment; filename="SheetJS.csv"');
-    // res.end(XLSX.utils.sheet_to_csv(wb.Sheets[wb.SheetNames[0]]));
-    // res.end(buffer);
+    res.end(this.fileManager.jsonToExcelBuffer(data, 'Utilizatori'));
   }
 
   @ApiParam({ name: 'id', type: Number })
