+++ backend/src/modules/user/user.module.ts
@@ -11,15 +11,14 @@ import { ApplicationModule } from '../application/application.module';
 import { UserHistory } from './entities/user-history.entity';
 import { UserApplicationService } from './services/user-application.service';
 import { MulterModule } from '@nestjs/platform-express';
+import { SharedModule } from 'src/shared/shared.module';
 
 @Module({
   imports: [
+    SharedModule,
     TypeOrmModule.forFeature([User, UserHistory]),
     forwardRef(() => OrganizationModule),
     ApplicationModule,
-    MulterModule.register({
-      dest: './download',
-    }),
   ],
   controllers: [ProfileController, AdminUserController],
   providers: [
