+++ backend/src/shared/controllers/shared.controller.ts
@@ -2,12 +2,12 @@ import { Controller, Get, Query } from '@nestjs/common';
 import { ApiQuery } from '@nestjs/swagger';
 import { SkipThrottle } from '@nestjs/throttler';
 import { Public } from 'src/common/decorators/public.decorator';
-import { FileManagerService } from '../services/file-manager.service';
+import { S3FileManagerService } from '../services/s3-file-manager.service';
 
 @Public()
 @Controller('')
 export class SharedController {
-  constructor(private readonly fileManagerService: FileManagerService) {}
+  constructor(private readonly fileManagerService: S3FileManagerService) {}
 
   @SkipThrottle()
   @Get('health')
